Changed tile component name and directory. Divided MovieList styles from SmallTile styled components. Small style updates.
