+++ src/pages/MovieList/MovieList.jsx
@@ -1,4 +1,4 @@
-import { TileSmall } from "./TileSmall";
+import { SmallTile } from "../../components/common/Tile/SmallTile/SmallTile";
 import {
   Button,
   Container,
@@ -11,47 +11,45 @@ import {
 
 export default function MovieList() {
   return (
-    <>
-      <Container>
-        <Content>
-          <Header>Popular movies</Header>
-          <TilesRow>
-            <Tile>
-              <TileSmall />
-            </Tile>
-            <Tile>
-              <TileSmall />
-            </Tile>
-            <Tile>
-              <TileSmall />
-            </Tile>
-            <Tile>
-              <TileSmall />
-            </Tile>
-            <Tile>
-              <TileSmall />
-            </Tile>
-            <Tile>
-              <TileSmall />
-            </Tile>
-            <Tile>
-              <TileSmall />
-            </Tile>
-            <Tile>
-              <TileSmall />
-            </Tile>
-          </TilesRow>
-        </Content>
-        <Pagination>
-          <Button>First</Button>
-          <Button>Previous</Button>
-          <span>
-            Page <strong>1</strong> of <strong>500</strong>
-          </span>
-          <Button>Next</Button>
-          <Button>Last</Button>
-        </Pagination>
-      </Container>
-    </>
+    <Container>
+      <Header>Popular movies</Header>
+      <Content>
+        <TilesRow>
+          <Tile>
+            <SmallTile />
+          </Tile>
+          <Tile>
+            <SmallTile />
+          </Tile>
+          <Tile>
+            <SmallTile />
+          </Tile>
+          <Tile>
+            <SmallTile />
+          </Tile>
+          <Tile>
+            <SmallTile />
+          </Tile>
+          <Tile>
+            <SmallTile />
+          </Tile>
+          <Tile>
+            <SmallTile />
+          </Tile>
+          <Tile>
+            <SmallTile />
+          </Tile>
+        </TilesRow>
+      </Content>
+      <Pagination>
+        <Button>First</Button>
+        <Button>Previous</Button>
+        <span>
+          Page <strong>1</strong> of <strong>500</strong>
+        </span>
+        <Button>Next</Button>
+        <Button>Last</Button>
+      </Pagination>
+    </Container>
   );
 }
